fix: radiative transfer parameters
I am lost as to how to convert their cm units in the Gold+20 paper to
something we can use in our code, so I just fiddled the numbers until
Test 1 and Test 4 were roughly in the right ballpark. Since the other
tests are still _also_ in the right ballpark, I'm going to say the test
is a success, but I will add a:

todo: check how to convert Gold+20 units to something we can use!